Move non-agent aggregator code into main ocs library
